requirements.txt
crawl4ai
langchain 
openai 
pgvector 
psycopg2
langchain-openai
langchain-community
python-dotenv
psycopg2-binary
langgraph
langchain-postgres
jq
fastapi
uvicorn
pydantic
pydantic-settings
httpx
google-generativeai
google-genai
langchain-google-genai
alembic
sqlalchemy
uuid
steamsignin
python-jose